Handle operations on deleted cards
When the hidden word position of a card is edited, the server
deletes the original card and replaces it with the new cards. As
a result, operations involving the original card can no longer be
performed. In those cases, the program should notify the user
where neccessary.

`CardType` has been replaced with `CardInterface` and
`AugmentedCardType` has been
replaced with an unexported
`AugmentedCard` class. Affected files have been edited.