Display answer in plain text after user's attempt
`Question` is now divided into `QuestionInput` and
`QuestionAnswer`. `QuestionInput` contains the
`input` element that the user interacts with, while
`QuestionAnswer` is displayed once the user has answered.

Since the answer is displayed in plain, it can be easily selected
with a double-click. Hence, Issue #24.

In the future, the displayed answer will be interactable,
allowing users to select which words to hide.