Install PyICU
In the Turkish language, the lowercase of I is not i but ı
(dotless i). Similary, the uppercase of i is İ (dotted I).

To correctly lower- and uppercase text, we'll perform
locale sensitive lower- and uppercasing.

JavaScript provides toLocaleLowerCase/toLocaleUpperCase
but Python doesn't have an equivalent. PyICU is our library of
choice which provides locale sensitive lower- and uppercasing
for Python.