Let user know if correct answer has been reached
This solves Issue #39.

The text color is red when a user begins inputing the answer; it
turns green when the text input box contains a correct answer.

This has been accomplished by introducing a new concept of
`inputPrelimStatus` which is either `correct` or `incorrect`
based on the real-time value of the contents of the input box.
`inputPrelimStatus` is updated each time the value in the input
box changes.