Tell user if the input is on the right track
As the user begins typing out the answer, the UI indicates whether
s/he is on the path to the right answer. If the input is correct, it
is highlighted in green, if it can be correct with some more input
(input is "app" and the correct answer is "apple"), it is
highlighted in gray, otherwise a red highlight is used.