Allow "duplicate" WordCards
Some databases can't distinguish between accented and
unaccented characters. For example, PythonAnywhere's
MySQL considers "warst" and "wärst" to be the same.

So that this don't cause a unique constraint violation when cards
are being prepared for a user of a particular list, in_bulk mode
was turned off in 1b28cd17acf1f2f4f4a94e9a1d8f59b5e913027f.

However, this leads to a very long waiting time when users try
to play a list for the first time. To avoid this, in_bulk mode has
been turned back on. To avoid the problems caused by this, the
unique constraint has been removed.